feat(history-service): remove history package type exports
In addition, the package and file structure has been adapted to that of the other packages.
