+++ packages/history-service/src/internal/browser-consumer-history.ts
