+++ packages/history-service/src/internal/static-consumer-history.ts
