+++ packages/history-service/src/internal/static-root-history.ts
